Fix Contracts dependency, add type conversion, update visual snapshots
Agent-Logs-Url: https://github.com/vormadal/Anything/sessions/a5774b60-b8f2-4208-92b3-ee8cf94a1a67

+++ src/Anything.Application/Features/ShoppingLists/Queries/GetShoppingLists.cs
@@ -29,7 +29,7 @@ public class GetShoppingListsHandler(IRepository<ShoppingList> listRepository, I
                     l.ModifiedOn,
                     l.DeletedOn,
                     items.Count(),
-                    l.Type))
+                    (int)l.Type))
             .ToListAsync(ct);
     }
 }
